move cases-per-day-card to cards folder
also, make the demo a bit more realistic
